+++ CHANGELOG.rst
@@ -31,3 +31,7 @@ Changelog
 - 1.1.1 ; Reorganized files and classes and made changes to documentation. Also added totally_not_a_rickroll() and pixelgrab(), and also mousepixelgrab().
 
 - 1.1.1.1 ; Minor bugfix to strhash()
+
+- 1.1.1.2 ; Added minor obfuscation functions
+
+- 1.1.1.3 ; Changed name of bhash() to bytehash()